requirements.txt
imageio>=2.6.1
ipython>=7.10.0
matplotlib>=3.1.0
numpy>=1.15.4
opencv>=3.4.2
pillow=6.2.1
python>=3.7.5
pytorch>=1.3.0
scikit-image>=0.16.2
scipy>=1.1.0
torchvision>=0.4.2
